# docs/swagger.yaml
basePath: /
definitions:
...
  dto.CreateSocialRequest:
    properties:
      name:
        maxLength: 191
        type: string
      social_media_url:
        maxLength: 191
        type: string
    required:
    - name
    - social_media_url
    type: object
  dto.CreateSocialResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      user_id:
        type: integer
    type: object
...
  dto.DeleteSocialResponse:
    properties:
      message:
        type: string
    type: object
  dto.DeleteUserResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.GetAllSocialResponse:
    properties:
      created_at:
        type: string
      id:
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
      user_id:
        type: integer
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        minLength: 6
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      token:
        type: string
    type: object
...
  dto.RegisterRequest:
    properties:
      age:
        maximum: 100
        minimum: 8
        type: integer
      email:
        maxLength: 191
        type: string
      password:
        minLength: 6
        type: string
      username:
        maxLength: 10
        type: string
    required:
    - age
    - email
    - password
    - username
    type: object
  dto.RegisterResponse:
    properties:
      age:
        type: integer
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
...
  dto.UpdateRequest:
    properties:
      email:
        maxLength: 191
        type: string
      username:
        maxLength: 10
        type: string
    required:
    - email
    - username
    type: object
  dto.UpdateResponse:
    properties:
      age:
        type: integer
      email:
        type: string
      id:
        type: integer
      updated_at:
        type: string
      username:
        type: string
    type: object
  dto.UpdateSocialRequest:
    properties:
      name:
        maxLength: 191
        type: string
      social_media_url:
        maxLength: 191
        type: string
    required:
    - name
    - social_media_url
    type: object
  dto.UpdateSocialResponse:
    properties:
      id:
        type: integer
      name:
        type: string
      social_media_url:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
  dto.UserResponse:
    properties:
      email:
        type: string
      id:
        type: integer
      username:
        type: string
    type: object
  helpers.ApiInfo:
    properties:
      message:
        type: string
      status:
        type: integer
    type: object
  helpers.ApiResponse:
    properties:
      data: {}
      info:
        $ref: '#/definitions/helpers.ApiInfo'
    type: object
host: localhost:8080
info:
  contact: {}
  title: MyGram Example API
  version: "1.0"
paths:
...
  /socialmedias/:
    get:
      description: Get all social
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.GetAllSocialResponse'
                  type: array
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/helpers.ApiResponse'
      security:
      - Authorization: []
      summary: Get all social
      tags:
      - Social
    post:
      consumes:
      - application/json
      description: Create new social
      parameters:
      - description: Create social body
        in: body
        name: createSocialRequest
        required: true
        schema:
          $ref: '#/definitions/dto.CreateSocialRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.CreateSocialResponse'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/helpers.ApiResponse'
      security:
      - Authorization: []
      summary: Create social
      tags:
      - Social
  /socialmedias/{socialMediaId}:
    delete:
      description: Delete social
      parameters:
      - description: Social media Id
        in: path
        name: socialMediaId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.DeleteSocialResponse'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/helpers.ApiResponse'
      security:
      - Authorization: []
      summary: Delete social
      tags:
      - Social
    put:
      consumes:
      - application/json
      description: Update social
      parameters:
      - description: Update social body
        in: body
        name: updateSocialRequest
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateSocialRequest'
      - description: Social media Id
        in: path
        name: socialMediaId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UpdateSocialResponse'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/helpers.ApiResponse'
      security:
      - Authorization: []
      summary: Update social
      tags:
      - Social
  /users/:
    delete:
      consumes:
      - application/json
      description: Delete User
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.DeleteUserResponse'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/helpers.ApiResponse'
      security:
      - Authorization: []
      summary: Delete User
      tags:
      - User
    put:
      consumes:
      - application/json
      description: Update User Data
      parameters:
      - description: Login body
        in: body
        name: registerRequest
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UpdateResponse'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/helpers.ApiResponse'
      security:
      - Authorization: []
      summary: Update User
      tags:
      - User
  /users/login:
    post:
      consumes:
      - application/json
      description: Login user
      parameters:
      - description: Login body
        in: body
        name: registerRequest
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/helpers.ApiResponse'
      summary: Login
      tags:
      - User
  /users/register:
    post:
      consumes:
      - application/json
      description: Register new user
      parameters:
      - description: Register body
        in: body
        name: registerRequest
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Success
          schema:
            allOf:
            - $ref: '#/definitions/helpers.ApiResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.RegisterResponse'
              type: object
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/helpers.ApiResponse'
      summary: Register User
      tags:
      - User
securityDefinitions:
  Authorization:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"